+++ StepTrackerTCA/Features/Records/ChildFeature/WorkoutsSummary/ChildFeature/Scores/ScoresView.swift
@@ -34,11 +34,9 @@ struct ScoresView: View {
                 .presentationDetents([.medium, .large])
         })
         .sheet(item: $store.scope(state: \.destination?.openGoal, action: \.destination.openGoal), content: { store in
-            //SetEditGoalView(store: store)
             WorkoutSubmissionView(store: store)
         })
         .sheet(item: $store.scope(state: \.destination?.openSubmitWorkout, action: \.destination.openSubmitWorkout), content: { store in
-            //SubmitWorkoutView(store: store)
             WorkoutSubmissionView(store: store)
         })
         .navigationDestination(
